Don't send typing notifications when the configuration changes
This was a bitch.  Our listeners get called when the UI is redrawn (orientation changes).
We do a few things.
1) don't send typing notifications if the # of characters changed before and after are 0
2) disable typing notifications while the UI is being drawn.  We do this by setting the lastChangedTime
to a sentinal value in onCreate and resetting it to a sane value in onResume.  onResume is called after onRestoreInstanceState
where we can retrieve any last set values for this field.