Refactor code, no functional changes
If we implement a single top activity, then Intent's sent to this Activity when the Activity
is already on the top will cause onNewIntent to get called on the existing Activity.  Confused?  I am.

Refactor the code so we can start doing this.  Bind all variables to views in onCreate, but initialize them
in separate functions.  Add listeners in onCreate so they're only called once.