Only allow our APK to be installed on internal storage
According to http://developer.android.com/guide/appendix/install-location.html,
external storage has several implications that make it unusable for a long-running App.  Most critically,
your application will not be able to receive the onboot intent since the system is booted before the environment
mounts the storage device (thus, it is unaware of your application).  It also means that the service for the app is killed
when the device enters mass storage mode.  As such, only allow our app to be installed on the internal storage.